[refactor] custom hook 제작 modal/afterGame #775
